Fix TBB download, verification, and extraction process
So after struggling to find a way to achieve idempotence using the ansible
unarchive module and then tar, I've decided it's not worth the time or effort.
Idempotence would require unarchiving the tarball in memory and computing the
hash of every file within it and then comparing each of those hashes to the hash
of the files on disk. That seems time/ computational intensive. Do see my
comment above the task "Extract TBB archive to the dir required by
tor-browser-crawler for this TBB version" for an explanation of the problems
with the unarchive module and `tar`'s `--diff` switch.

Signed-off-by: Noah Vesely <noah@freedom.press>