Detect captcha in page, change in dir name
CrawlJob is informed of the captcha. It renames the corresponding
directory and annotates that it is a captcha, so that its "path"
property maps to the correct renamed directory if appropriate

We should refine the reaction to a captcha page, indeed. This is a first
step, that for me it works for the google recaptcha engine used by
Cloudflare.

Added a test for this functionality, that checks that there are two
directory beginning with 'captcha_' in a 1-batch 2-visits crawl of
https://cloudflare.com.

Since my machine does not have 'eth0', I've added a library (netifaces)
that allows me to select an interface connected to the internet. We
could use this in the future, to improve the portability across
different Linux distros/OSes. So far, I've added a command line argument
to specify a concrete interface by hand. By default, the device is
'eth0' but now you have to tell the Sniffer from which device you want
to sniff.

Regarding robustness, in the try/finally block of pytbcrawler I make
sure both the driver and the controller quit before exiting. This avoids
the problem of having a tor instance running if an exception happens
inside the code or if the user interrupts the execution using the
keyboard. I've tested it in the test_cloudflare_captcha_page, but not in
a "real" crawl.